requirements.txt
numpy
torch>=1.0
torchvision

matplotlib
tensorflow
tensorboard
terminaltables
pillow
tqdm
